requirements.txt
discord.py-self
openai
python-dotenv
pillow
aiohttp
